implemented suggestions (#207)
